eina_stringshare: cast in order to fix a warning
Summary:
casting here is safe, as before we check and return on size < 1.
Depends on D6379

Reviewers: devilhorns, Hermet

Reviewed By: Hermet

Subscribers: cedric, #committers, zmike

Tags: #efl

Differential Revision: https://phab.enlightenment.org/D6380

+++ src/lib/eina/eina_stringshare.c
@@ -679,7 +679,7 @@ eina_stringshare_nprintf(unsigned int len, const char *fmt, ...)
 
    if (size < 1)
      return "";
-   if (size > len)
+   if ((unsigned int)size > len)
      size = len;
 
    return eina_stringshare_add_length(tmp, size);
